Add files via upload
Here I am adding the css code that i have made changes to. It has been a bit of a difficult process adding my files through the git commands in terminal, so I have been making changes locally, tracking my progress using the live server extension on VSCode. I have jut figured out you can use a drag and drop method to add your files to your GitHub repository.

In this Commit, I have made changes to the existing CSS that improves the functionality in various ways. Firstly, there was a broken link that was fixed. Additionally, changes have been made to improve the way the website responds to resizing. Now, text and images stay in the containers upon resizing. Furthermore, comments have been provided explaining the changes that were made. Also, some formatting was changed to reduce some unnecessary code as well, and to follow better conventions. 

This was also the case in the html file. I changed the unordered list to a nav element, added an id element to make the anchor tag function as it was supposed to, and provided some comments explaining all the changes that were made.